Do not include rogue-dev in the Yocto images.
`rogue-dev` appears to somehow pull along a huge amount of dependencies
inflating the generated image.
Apparently this should also have been included in the images produced by
the old Petalinux bold process, but inspecting build files suggests that
the setting somehow ended up ignored, so the `rogue-dev` never was in
the images.